fix: nodrag option: X browser support.
The `path` method appears to only be supported by Chrome. composedPath()
seems to be an alternative, but it's not supported by IE.

fixed by checking the element which incurred the mousedown event as well
as it's parent for the presence of the 'nodrag' attribute will suffice
for where it's being used at present.